requirements.txt
praw
requests
beautifulsoup4
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
snscrape
python-dotenv
smtplib
email
